--- a/migrations/20170312165531_index_dependencies/down.sql
+++ b/migrations/20170312165531_index_dependencies/down.sql
@@ -1,0 +1,1 @@
+DROP INDEX dependencies_crate_id_version_id_idx;