--- a/migrations/20170312165531_index_dependencies/up.sql
+++ b/migrations/20170312165531_index_dependencies/up.sql
@@ -1,0 +1,1 @@
+CREATE INDEX ON dependencies (crate_id, version_id);